Users/DELL/Desktop/mybook/chatbot/main.py
import os
import time
import uvicorn
from google import genai
from dotenv import load_dotenv
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from document_processor import read_markdown_files, chunk_content
from embedding_generator import EmbeddingGenerator
from vector_store import VectorStore
from typing import List, Optional

# Load environment variables
load_dotenv()

client = genai.Client(api_key=os.getenv("GEMINI_API_KEY"))

QDRANT_URL = os.getenv("QDRANT_URL")
QDRANT_API_KEY = os.getenv("QDRANT_API_KEY")
QDRANT_COLLECTION_NAME = os.getenv("QDRANT_COLLECTION_NAME")
CHAT_MODEL = os.getenv("CHAT_MODEL", "gemini-2.0-flash")

if not all([QDRANT_URL, QDRANT_API_KEY, QDRANT_COLLECTION_NAME]):
    raise ValueError("Missing one or more environment variables.")

# Initialize FastAPI app
app = FastAPI()

app.add_middleware(
    CORSMiddleware,
    allow_origins=["http://localhost:3000", "*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Initialize our components
embedding_generator = EmbeddingGenerator()
vector_store = VectorStore()

class ChatRequest(BaseModel):
    question: str
    selected_text: Optional[str] = None

class ChatResponse(BaseModel):
    answer: str
    sources: List[str]

class IngestResponse(BaseModel):
    message: str

def ingest_documents():
    docs_path = "../book/docs/"
    documents = read_markdown_files(docs_path)
    chunks = chunk_content(documents)

    contents = [chunk["content"] for chunk in chunks]
    metadatas = [{"filepath": chunk["filepath"], "start_token": chunk["start_token"], "end_token": chunk["end_token"]} for chunk in chunks]

    embeddings = embedding_generator.generate_embeddings(contents)

    if not embeddings:
        raise Exception("Failed to generate embeddings for documents.")

    vector_store.recreate_collection()
    vector_store.upsert_vectors(contents, embeddings, metadatas)

@app.post("/ingest", response_model=IngestResponse)
async def ingest_content():
    try:
        ingest_documents()
        return IngestResponse(message="Content ingested successfully.")
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

def create_embedding(text: str) -> List[float]:
    embeddings = embedding_generator.generate_embeddings([text])
    if embeddings:
        return embeddings[0]
    raise Exception("Failed to generate embedding.")

@app.post("/chat", response_model=ChatResponse)
async def chat_with_bot(request: ChatRequest):
    query_text = request.question
    if request.selected_text:
        query_text = f"Context: {request.selected_text}\nQuestion: {request.question}"

    query_embedding = create_embedding(query_text)
    search_result = vector_store.search_vectors(query_embedding)

    context_chunks = [hit["content"] for hit in search_result]
    sources = [f"{hit.get('filepath', '')}#L{hit.get('start_token', '')}-L{hit.get('end_token', '')}" for hit in search_result]

    context_str = "\n\n".join(context_chunks)
    time.sleep(3)      

    response = client.models.generate_content(
        model=CHAT_MODEL,
        contents=f"You are a helpful AI assistant for a Physical AI Robotics textbook. Only answer questions related to the book content. If someone greets you, greet back briefly. Keep answers concise and clear.\n\nContext: {context_str}\n\nQuestion: {request.question}"
    )

    answer = response.text
    return ChatResponse(answer=answer, sources=sources)

if __name__ == "__main__":
    uvicorn.run(app, host="0.0.0.0", port=8000)