Users/DELL/Desktop/mybook/chatbot/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
qdrant-client==1.8.0
python-dotenv==1.0.1
markdown==3.6
tiktoken==0.6.0
google-genai